move `ipython notebook trust` to `ipython trust`
and move App definition to nbformat.sign (maybe it should get its own file).
